Rename command abstraction to shell executor
