+++ www/tests/Unit/ShellExecutorTest.php
@@ -4,18 +4,18 @@ declare(strict_types=1);
 
 namespace Tests;
 
-use FalconChristmas\Fpp\Command\CommandExecutor;
-use FalconChristmas\Fpp\Command\CommandResult;
+use FalconChristmas\Fpp\Shell\ShellExecutor;
+use FalconChristmas\Fpp\Shell\ShellResult;
 use PHPUnit\Framework\TestCase;
 
-final class CommandExecutorTest extends TestCase
+final class ShellExecutorTest extends TestCase
 {
     public function testRunReturnsSuccessfulResult(): void
     {
-        $executor = new CommandExecutor();
+        $executor = new ShellExecutor();
         $result = $executor->run('echo success');
 
-        $this->assertInstanceOf(CommandResult::class, $result);
+        $this->assertInstanceOf(ShellResult::class, $result);
         $this->assertSame('echo success 2>&1', $result->getCommand());
         $this->assertSame(['success'], $result->getOutput());
         $this->assertSame(0, $result->getExitCode());
@@ -25,10 +25,10 @@ final class CommandExecutorTest extends TestCase
 
     public function testRunWithoutRedirectStdErr(): void
     {
-        $executor = new CommandExecutor();
+        $executor = new ShellExecutor();
         $result = $executor->run('echo success', false);
 
-        $this->assertInstanceOf(CommandResult::class, $result);
+        $this->assertInstanceOf(ShellResult::class, $result);
         $this->assertSame('echo success', $result->getCommand());
         $this->assertSame(['success'], $result->getOutput());
         $this->assertSame(0, $result->getExitCode());
@@ -38,7 +38,7 @@ final class CommandExecutorTest extends TestCase
 
     public function testRunCapturesFailureExitCode(): void
     {
-        $executor = new CommandExecutor();
+        $executor = new ShellExecutor();
         $result = $executor->run('php -r "exit(3);"');
 
         $this->assertSame(3, $result->getExitCode());
